apply thought color to superscript
