+++ src/components/Bullet.tsx
@@ -15,6 +15,7 @@ import { isMac, isSafari, isTouch, isiPhone } from '../browser'
 import attributeEquals from '../selectors/attributeEquals'
 import findDescendant from '../selectors/findDescendant'
 import { getAllChildrenAsThoughts, getChildren } from '../selectors/getChildren'
+import getElementFill from '../selectors/getElementFill'
 import getLexeme from '../selectors/getLexeme'
 import getThoughtById from '../selectors/getThoughtById'
 import isContextViewActive from '../selectors/isContextViewActive'
@@ -45,8 +46,6 @@ interface BulletProps {
 }
 
 const isIOSSafari = isTouch && isiPhone && isSafari()
-const styleRegex = /style="[^"]*(background-)?color:\s*([^;"'>]+)[^"]*"/i
-const fontColorRegex = /<font[^>]*color="([^"]+)"[^>]*>/i
 
 const glyph = cva({
   base: {
@@ -441,76 +440,7 @@ const Bullet = ({
     return children.length < Object.keys(thought.childrenMap).length
   })
 
-  /** Check if the entire value is wrapped in a single font or span tag. */
-  const isWrappedInSingleTag = (str: string): boolean => {
-    // Check if there's any text before the first tag or after the last tag
-    const firstTagIndex = str.indexOf('<')
-    const lastTagIndex = str.lastIndexOf('>')
-
-    if (firstTagIndex > 0 || lastTagIndex < str.length - 1) {
-      return false
-    }
-
-    const tagStack: string[] = []
-    let currentPos = 0
-
-    while (currentPos < str.length) {
-      // Find next tag
-      const openIndex = str.indexOf('<', currentPos)
-      if (openIndex === -1) break
-
-      const closeIndex = str.indexOf('>', openIndex)
-      if (closeIndex === -1) return false
-
-      const tag = str.substring(openIndex, closeIndex + 1)
-      currentPos = closeIndex + 1
-
-      if (tag.startsWith('</')) {
-        // Closing tag
-        const tagName = tag.slice(2, -1).toLowerCase()
-        // If stack is empty or last opening tag doesn't match, return false
-        if (tagStack.length === 0 || tagStack[tagStack.length - 1] !== tagName) {
-          return false
-        }
-
-        // Remove matching opening tag
-        tagStack.pop()
-
-        // If stack is empty before end of string, only valid if this is the last tag
-        if (tagStack.length === 0 && currentPos <= str.length) {
-          const remainingTags = str.slice(currentPos).match(/<\/?[a-z]+/g)
-          return !remainingTags
-        }
-      } else if (!tag.endsWith('/>')) {
-        // Opening tag (excluding self-closing tags)
-        const tagName = tag.slice(1).split(/[\s>]/)[0].toLowerCase()
-        tagStack.push(tagName)
-      }
-    }
-
-    // Valid if exactly one tag remains in stack (the outer tag)
-    return tagStack.length === 1 && (tagStack[0] === 'font' || tagStack[0] === 'span')
-  }
-
-  const fill = useSelector(state => {
-    const thought = getThoughtById(state, thoughtId)
-    if (!thought) return undefined
-
-    const value = thought.value
-    // Check if the entire value is wrapped in a single font or span tag
-    const isWrappedInTag = isWrappedInSingleTag(value)
-    if (isWrappedInTag) {
-      // Check for background-color and color in style attribute
-      const styleMatch = value.match(styleRegex)
-      if (styleMatch) return styleMatch[2]
-
-      // If no background-color, check for font color
-      const fontColorMatch = value.match(fontColorRegex)
-      if (fontColorMatch) return fontColorMatch[1]
-    }
-
-    return undefined
-  })
+  const fill = useSelector(state => getElementFill(state, thoughtId!))
 
   const isExpanded = useSelector(state => !!state.expanded[hashPath(path)])
   const isBulletExpanded = isCursorParent || isCursorGrandparent || isEditing || isExpanded
